# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CSVReportRequest": {
        "properties": {
          "res_dir": {
            "title": "Res Dir",
            "type": "string"
          },
          "ro5": {
            "$ref": "#/components/schemas/Ro5Thresholds",
            "default": {
              "hba": 10,
              "hbd": 5,
              "logp": 5.0,
              "max_violations": 1,
              "mw": 500.0
            }
          },
          "src_dir": {
            "default": "",
            "title": "Src Dir",
            "type": "string"
          },
          "top_n": {
            "default": 10,
            "title": "Top N",
            "type": "integer"
          }
        },
        "required": [
          "res_dir"
        ],
        "title": "CSVReportRequest",
        "type": "object"
      },
      "CSVReportResponse": {
        "properties": {
          "csv_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Csv Path"
          },
          "rows": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CSVRow"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "title": "CSVReportResponse",
        "type": "object"
      },
      "CSVRow": {
        "properties": {
          "hba": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hba"
          },
          "hbd": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hbd"
          },
          "ligand": {
            "title": "Ligand",
            "type": "string"
          },
          "logp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logp"
          },
          "mw": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mw"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "rule_of_5": {
            "default": "Unknown",
            "title": "Rule Of 5",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "tpsa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tpsa"
          }
        },
        "required": [
          "rank",
          "ligand",
          "score"
        ],
        "title": "CSVRow",
        "type": "object"
      },
      "Candidate": {
        "properties": {
          "dir": {
            "title": "Dir",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "name",
          "score",
          "dir"
        ],
        "title": "Candidate",
        "type": "object"
      },
...
      "ConsensusRequest": {
        "properties": {
          "results_dir": {
            "title": "Results Dir",
            "type": "string"
          },
          "src_dir": {
            "default": "",
            "title": "Src Dir",
            "type": "string"
          }
        },
        "required": [
          "results_dir"
        ],
        "title": "ConsensusRequest",
        "type": "object"
      },
      "ConsensusResponse": {
        "properties": {
          "csv_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Csv Path"
          },
          "results": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ConsensusResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "title": "ConsensusResponse",
        "type": "object"
      },
      "ConsensusResult": {
        "properties": {
          "consensus_rank": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consensus Rank"
          },
          "contact_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Rank"
          },
          "contact_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Score"
          },
          "energy_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Energy Rank"
          },
          "ligand": {
            "title": "Ligand",
            "type": "string"
          },
          "mmff_energy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mmff Energy"
          },
          "vina_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vina Rank"
          },
          "vina_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vina Score"
          }
        },
        "required": [
          "ligand"
        ],
        "title": "ConsensusResult",
        "type": "object"
      },
...
      "ExportTopRequest": {
        "properties": {
          "results_dir": {
            "title": "Results Dir",
            "type": "string"
          },
          "src_dir": {
            "title": "Src Dir",
            "type": "string"
          },
          "top_n": {
            "default": 10,
            "title": "Top N",
            "type": "integer"
          }
        },
        "required": [
          "src_dir",
          "results_dir"
        ],
        "title": "ExportTopRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IUPACResponse": {
        "properties": {
          "common_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Common Name"
          },
          "iupac_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iupac Name"
          },
          "smiles": {
            "title": "Smiles",
            "type": "string"
          }
        },
        "required": [
          "smiles"
        ],
        "title": "IUPACResponse",
        "type": "object"
      },
...
      "LoadResultsResponse": {
        "properties": {
          "candidates": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Candidate"
            },
            "title": "Candidates",
            "type": "array"
          }
        },
        "title": "LoadResultsResponse",
        "type": "object"
      },
...
      "Ro5Thresholds": {
        "properties": {
          "hba": {
            "default": 10,
            "title": "Hba",
            "type": "integer"
          },
          "hbd": {
            "default": 5,
            "title": "Hbd",
            "type": "integer"
          },
          "logp": {
            "default": 5,
            "title": "Logp",
            "type": "number"
          },
          "max_violations": {
            "default": 1,
            "title": "Max Violations",
            "type": "integer"
          },
          "mw": {
            "default": 500,
            "title": "Mw",
            "type": "number"
          }
        },
        "title": "Ro5Thresholds",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Drug Discovery Suite — REST + WebSocket API",
    "title": "Helix Core Backend",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/results/consensus": {
      "post": {
        "description": "Multi-method consensus ranking: Vina score + MMFF energy + contact count.",
        "operationId": "consensus_scoring_api_results_consensus_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsensusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsensusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Consensus Scoring",
        "tags": [
          "Results Explorer"
        ]
      }
    },
    "/api/results/csv-report": {
      "post": {
        "description": "Generate a CSV report with docking scores + ADMET properties.",
        "operationId": "csv_report_api_results_csv_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CSVReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CSVReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Csv Report",
        "tags": [
          "Results Explorer"
        ]
      }
    },
    "/api/results/export-sdf": {
      "post": {
        "description": "Convert a list of compounds from a source directory to a combined SDF file.\n\nBody: { \"names\": [\"mol1\", \"mol2\"], \"src_dir\": \"path\" }\nReturns: { \"sdf_path\": \"...\", \"count\": N }",
        "operationId": "export_sdf_api_results_export_sdf_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Req",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Sdf",
        "tags": [
          "Results Explorer"
        ]
      }
    },
    "/api/results/export-top": {
      "post": {
        "description": "Export top N docked results + original ligands into separate folders.",
        "operationId": "export_top_hits_api_results_export_top_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportTopRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Top Hits",
        "tags": [
          "Results Explorer"
        ]
      }
    },
    "/api/results/iupac": {
      "get": {
        "description": "Resolve SMILES to IUPAC/common name via PubChem.",
        "operationId": "resolve_iupac_api_results_iupac_get",
        "parameters": [
          {
            "in": "query",
            "name": "smiles",
            "required": true,
            "schema": {
              "title": "Smiles",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IUPACResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Iupac",
        "tags": [
          "Results Explorer"
        ]
      }
    },
    "/api/results/load": {
      "get": {
        "description": "Scan a results directory and return ranked candidates.",
        "operationId": "load_results_api_results_load_get",
        "parameters": [
          {
            "in": "query",
            "name": "dir",
            "required": true,
            "schema": {
              "title": "Dir",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadResultsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Load Results",
        "tags": [
          "Results Explorer"
        ]
      }
    },
    "/api/results/load-csv": {
      "get": {
        "description": "Load and parse an existing Report CSV for charting.",
        "operationId": "load_csv_api_results_load_csv_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CSVReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Load Csv",
        "tags": [
          "Results Explorer"
        ]
      }
    },
...
  }
}